removed dynamic routing for product categories
