--- inlang/source-code/website/src/pages/@product/index.page.server.tsx
@@ -1,23 +1,0 @@
-
-import { render } from "vite-plugin-ssr/abort"
-import { products } from "./products.jsx"
-
-export async function onBeforeRender(pageContext: any) {
-  const { product } = pageContext.routeParams
-
-  const page = products.find((page) => page.slug === product)
-
-  if (!page) {
-		if (product === "documentation") return
-
-		throw render(404)
-	}
-
-  return {
-    pageContext: {
-      pageProps: {
-        ...page,
-      },
-    },
-  }
-}
