--- inlang/source-code/website/src/pages/@product/products.tsx
@@ -1,37 +1,0 @@
-export const products = [
-	{
-		slug: "application",
-		content: {
-			title: "Global Application",
-			description: "Globalization infrastructure for software",
-		},
-	},
-	{
-		slug: "email",
-		content: {
-			title: "Global Email",
-			description: "Stay in touch with your customers in their native language.",
-		},
-	},
-	{
-		slug: "payment",
-		content: {
-			title: "Global Payment",
-			description: "Enable your customers to pay with ease using their favorite method.",
-		},
-	},
-	{
-		slug: "website",
-		content: {
-			title: "Global Website",
-			description: "Let your website speak the language of your customers.",
-		},
-	},
-	{
-		slug: "document",
-		content: {
-			title: "Global Document",
-			description: "Documents made to collaborate with version control and multi-language support.",
-		},
-	},
-]
