+++ inlang/source-code/website/src/pages/application/index.page.server.tsx
@@ -1,0 +1,17 @@
+const page = {
+	slug: "application",
+	content: {
+		title: "Global Application",
+		description: "Globalization infrastructure for software",
+	},
+}
+
+export async function onBeforeRender() {
+	return {
+		pageContext: {
+			pageProps: {
+				...page,
+			},
+		},
+	}
+}
