+++ inlang/source-code/website/src/pages/email/index.page.server.tsx
@@ -1,0 +1,17 @@
+const page = {
+	slug: "email",
+	content: {
+		title: "Global Email",
+		description: "Stay in touch with your customers in their native language.",
+	},
+}
+
+export async function onBeforeRender() {
+	return {
+		pageContext: {
+			pageProps: {
+				...page,
+			},
+		},
+	}
+}
