+++ inlang/source-code/website/src/pages/payment/index.page.server.tsx
@@ -1,0 +1,17 @@
+const page = {
+	slug: "payment",
+	content: {
+		title: "Global Payment",
+		description: "Enable your customers to pay with ease using their favorite method.",
+	},
+}
+
+export async function onBeforeRender() {
+	return {
+		pageContext: {
+			pageProps: {
+				...page,
+			},
+		},
+	}
+}
