+++ inlang/source-code/website/src/pages/payment/index.page.tsx
@@ -1,0 +1,5 @@
+import { CategoryPage, type PageProps } from "../Category.jsx"
+
+export function Page(props: PageProps) {
+	return <CategoryPage slug={props.slug} content={props.content} />
+}
