+++ inlang/source-code/website/src/pages/website/index.page.server.tsx
@@ -1,0 +1,17 @@
+const page = {
+	slug: "website",
+	content: {
+		title: "Global Website",
+		description: "Let your website speak the language of your customers.",
+	},
+}
+
+export async function onBeforeRender() {
+	return {
+		pageContext: {
+			pageProps: {
+				...page,
+			},
+		},
+	}
+}
